Add file base persistent backend for auto merge queue.
By this change, this application uses `$XDG_CONFIG_HOME/popuko` as
the config dir by default.
You can change this path with
`-config-base-dir` option. If you specify
`config-base-dir=$HOME/Downloads/`, this application uses
`$HOME/Downloads/popuko` as the config dir.a

By this change, this application stores the auto-merge queue
as json file to `/queue` dir in the config dir.