Create the backup file on writing a queue to the file.
If the application fails down when it is writing to the file,
the content of its file may be wrecked.

To resolve it, we use the following steps when we write a content to the
file:

1. Creare the backup file by renaming the file to `***.old`.
2. Write a content to the file.
3. Remove the file created in steps 1).

As alternative approach, we could implement a graceful shutdown for this.
But we thought it is hard to implement it correctly. So we don't accept
its approach at this moment.